Fix AIO Builds
This is a companion code drop related to the os-services commit
with the same title.   It provides a genesis style inventory for
AIO configurations.  This project does not use this new capability
as the scripting logic uses a function to determine whether this
is a real genesis configuration.  In this way, the ceph scripts
avoid references to /var/oprc/inventory.yml in the AIO environment.
This project has non-overlapping code paths for AIO and genesis.
Opsmgr does not.  It has playbooks which assume the genesis inventory
file, so this change provides a vehicle for providing that data
in an AIO environment.  This code is impacted as there is a
common script which applies to all projects.

Change-Id: I6e029c87993b1dc283c7c623d0d4c2961e40fc2d
(cherry picked from commit 96ae3326a00ceeeee05c3454b2d9fcd024bae83c)