Move Ceph disk prep to create-cluster
Move the Ceph disk prep step to the create-cluster call.  This
move allows the caller to modify the disk list in the inventory
between the bootstrap-cluster and create-cluster step, which is
usually when they first have a chance to log into the provisioned
deployment targets and see the disk names.

Change-Id: Ia0e71909b809cf06f9e9edda74f716ed968d5c47