Fix /var/log/ceph/ceph-osd.0.log permissions
Ceph ansible deployments use the wrong file permissions
when creating the file /var/log/ceph/ceph-osd.0.log.
Within "ceph-disk prepare" the code mistakenly creates
that file with root.root permissions.  This prevents the
real ceph osd with id 0 from creating its own log file,
since that process runs as "ceph" and the file is owned
by "root".

This is technically a bug in ceph-disk [2] not in
ceph-ansible [1] but we can use ceph-ansible to
work around the issue for versions of ceph which don't
have the fix.

Since ceph-ansible supports versions of ceph where
the ceph user is not present, we only want to chown
the file to "ceph.ceph" if the ceph user exists.  It
is also necessary to signal the the osd 0 process so
that it retries the open on the file now that it has
the proper permissions.

This drop also adds mkdiffs support to ceph-services, so that
it is easier to patch ceph-ansible.

[1] https://github.com/ceph/ceph-ansible/issues/1223
[2] http://tracker.ceph.com/issues/18538

Change-Id: Icbce7cf90ff6f08b5ff8c896c625d9b2fd6a4409