ceph disk prep: Add sanity check for missing disks
Add a sanity check to verify that all the specified
journal-devices and osd-devices are present on each
ceph osd host.

If the sanity check fails on all hosts, the playbook
will end immediately.

If the sanity check fails on some hosts but not others,
the playbook should run to completion performing ceph
disk prep on any sane hosts.  The calling script
(create-cluster-ceph.sh) will recognize the failure
and stop running.

In the runtime output from the play, the error message
"One of the specified disks is not present" will be
displayed.  The specific detail on which host/device
is missing will be displayed in the task just above the
sanity check error msg.

To skip the sanity check, set "skip_missing_disk_sanity: true"
in playbooks/roles/ceph_hosts/defaults/main.yml.

Change-Id: Ic8819eb7c7b2590802b7c73b91f2e7240d2050ea