Disable predeploy playbook call
Disable the call to the predeploy playbook.
When create-cluster-ceph is recalled on a cluster that is
up and successful the ceph disk preparation will destroy
cluster data but ceph retains mounts.  This then causes
ceph-ansible to fail.

Disable the automatic calling of this playbook until it can
gracefully do no harm on re-entrance.

Change-Id: I5cfca379c54b5fb8f9c97d9ea01c697e3cf7f7d2