Remove Ubuntu Cloud Archive Use
Canonical is not yet shipping Ceph for Xenial in
the Ubuntu Cloud Archive (UCA).  We used UCA for Ceph for
Trusty because it contained newer versions. The use of UCA is
causing occasional failures with unauthenticated packages so we
will stop using that Apt repo until Canonical ships different
versions of Ceph for Xenial in UCA than it ships in xenial-updates.

Change-Id: Ic93e5bf8491b118dcde980ccaa5bbd0849457434