Avoid ceph disk prep on any active ceph host
If this is a freshly installed host where ceph is not
configured, we want to clear the disks.  On the other
hand, if ceph is configured we don't want to clear
the disks.   This way the playbook (and any callers)
can be used in a reentrant manner that does no harm
if it is called multiple times.

We can't perform a check on a per device basis
because journal-devices are accessed as raw partitions
(not mounted).  So we check for mounted ceph osd devices
on a per host basis instead.

Change-Id: Ic1caf52fe360a98004642c369ad0d6a25dcc4b9b