Do not display shortcut inventory stanzas
The ceph_osd_shortcut contains the inventory stanza for
ceph_osd nodes.  The ds_shortcut contains the 'domain-settings'
for those nodes.  Neither of these should be displayed, since
they can potentially contain secretive information (such as
passwords) that should not be printed out at runtime.  The
ceph journal and osd device lists are really all that need
to be printed.

Change-Id: I327d91629a2dd6e9e0cf68d38a0d95ec613a977f