Separate role from template name
Make use of new values in the cluster-genesis inventory to separate
the role of the node from the name of the node template.

This will let us have nodes that are not in the controllers node
template be ceph-monitors, which lets us converge nodes.

Change-Id: I18683d3bdc17995a8bf05296cd4a4c9ffacb6046
(cherry picked from commit 55bb9eea180b55d8fc27a1e71b632b5bc40cb56b)